Move CreateTraceLocation DI logic into sub-package
