+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/environment/eventregistration/createtracelocation/TraceLocationCDNHttpClient.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.environment.eventregistration
+package de.rki.coronawarnapp.environment.eventregistration.createtracelocation
 
 import javax.inject.Qualifier
