+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/EventRegistrationModule.kt
@@ -2,7 +2,7 @@ package de.rki.coronawarnapp.eventregistration
 
 import dagger.Binds
 import dagger.Module
-import de.rki.coronawarnapp.environment.eventregistration.CreateTraceLocationModule
+import de.rki.coronawarnapp.environment.eventregistration.createtracelocation.CreateTraceLocationModule
 import de.rki.coronawarnapp.eventregistration.checkins.download.DownloadedCheckInsRepo
 import de.rki.coronawarnapp.eventregistration.checkins.download.FakeDownloadedCheckInsRepo
 import de.rki.coronawarnapp.eventregistration.storage.repo.DefaultTraceLocationRepository
